+++ src/modules.mjs
@@ -77,7 +77,7 @@ function calculateLibraries() {
     libraries.push('libmemoryprofiler.js');
   }
 
-  if (SUPPORT_BASE64_EMBEDDING) {
+  if (SUPPORT_BASE64_EMBEDDING || ENVIRONMENT_MAY_BE_SHELL) {
     libraries.push('libbase64.js');
   }